Refactor code to fix bugs
