+++ game.rb
@@ -1,5 +1,5 @@
 require 'date'
-require_relative './item'
+require_relative 'item'
 
 class Game < Item
   attr_accessor :multiplayer, :last_played_at, :publish_date
