+++ game_methods.rb
@@ -1,5 +1,5 @@
-require_relative './app'
-require_relative './game'
+require_relative 'app'
+require_relative 'game'
 
 def show_games
   if @game_list.empty?
@@ -14,17 +14,26 @@ def show_games
 end
 
 def create_game
-  print 'Multiplayer [y/n]: '
-  multiplayer = gets.chomp
-  print 'Last played at [yyyy/mm/dd]: '
+  print 'Multiplayer [Y/N]: '
+  multiplayer = gets.chomp.split.map(&:capitalize).join(' ')
+  print 'Last played at [dd/mm/yyyy]: '
   last_played_at = gets.chomp
-  puts 'Please state when the game was published in this format - [yyyy/mm/dd]:'
+  if last_played_at =~ /^\d{2}\/\d{2}\/\d{4}$/
+  puts 'Please state when the game was published in this format - [dd/mm/yyyy]:'
+  else
+    puts "Invalid date format. Please enter date in dd/mm/yyyy format."
+  end
+  
   publish_date = gets.chomp
+  if publish_date =~ /^\d{2}\/\d{2}\/\d{4}$/
   new_game = Game.new(multiplayer, last_played_at, publish_date)
   @game_list << new_game
   create_dir
   save_files
   print 'Game has been added succesfully'
+  else
+    puts "Invalid date format. Please enter date in dd/mm/yyyy format."
+  end
 end
 
 def add_game(multiplayer, last_played_at, publish_date)
