+++ main.rb
@@ -1,12 +1,13 @@
-require_relative './app'
-require_relative './menu'
-require_relative './game_file'
-require_relative './book_data'
+require_relative 'app'
+require_relative 'menu'
+require_relative 'game_data'
+require_relative 'book_data'
 
 def main
   app = App.new
-  Dir.exist?('./storage') ? app.read_files : nil
-  Dir.exist?('./storage') ? app.read_files2 : nil
+  app.read_files if Dir.exist?('./storage')
+  app.read_files2 if Dir.exist?('./storage')
+
   loop do
     menu(app)
   end
