+++ musicalbum.rb
@@ -1,5 +1,5 @@
-require_relative './app'
-require_relative './item'
+require_relative 'app'
+require_relative 'item'
 require 'date'
 
 class MusicAlbum < Item
@@ -37,15 +37,20 @@ def list_music
 end
 
 def create_music
-  puts 'Is it on spotify? [y/n]: '
-  on_spotify = gets.chomp
+  puts 'Is it on spotify? [Y/N]: '
+  on_spotify = gets.chomp.split.map(&:capitalize).join(' ')
   puts 'Please state when was the music published [dd/mm/yyyy]:'
   publish_date = gets.chomp
-  new_music = MusicAlbum.new(on_spotify, publish_date)
-  @music_list.push(new_music)
-  create_dir
-  save_files
-  print 'Music Album Added Succesfully'
+  if publish_date =~ %r{^\d{2}/\d{2}/\d{4}$}
+    new_music = MusicAlbum.new(on_spotify, publish_date)
+    @music_list.push(new_music)
+    create_dir
+    save_files
+    print 'Music Album Added Succesfully'
+    
+  else
+    puts 'Invalid date format. Please enter date in dd/mm/yyyy format.'
+  end
 end
 
 def add_music(on_spotify, publish_date)
