+++ storage/music.json
@@ -1,1 +1,1 @@
-[{"ref":"#<MusicAlbum:0x000000011508fd20>","value":{"id":83,"genre":null,"author":null,"label":null,"publish_date":"2023/03/03","archived":false,"on_spotify":false}}]
+[{"ref":"#<MusicAlbum:0x000002a7b9076af8>","value":{"id":58,"genre":null,"author":null,"label":null,"publish_date":"12/12/2020","archived":false,"on_spotify":false}}]
